adds  `chunkModuls: true` to toJson options
prior to webpack v3.0, Stats.toJson's option had the `chunkModules` property with default value of false. (after v3.0 its default value is true). SunBurstChart needs the modules property for the chuncks for rendering the chart, therefor in the toJson options the `chunkModules: true` property was added to allow rendering the chart for older webpack versions.